migrate create entry tests
